+++ tests/test_core.py
@@ -14,7 +14,7 @@ from rule_estimator import *
 @pytest.fixture
 def model():
     return RuleClassifier(
-        LesserThanNode("petal length (cm)", 1.91, # BinaryNode
+        LesserThanSplit("petal length (cm)", 1.91, # BinarySplit
             if_true=PredictionRule(prediction=0), # PredictionRule: always predict 0
             if_false=CaseWhen([
                         # Go through these rules and if one applies, assign the prediction
@@ -30,7 +30,7 @@ def model():
 @pytest.fixture
 def model_with_estimator():
     return RuleClassifier(
-        LesserThanNode("petal length (cm)", 1.9, 
+        LesserThanSplit("petal length (cm)", 1.9, 
                 if_true=PredictionRule(0), 
                 if_false=CaseWhen([
                         LesserThan("petal length (cm)", 4.5, 1),